Don't pass api_client into PanopticonRegisterer
The API client and the way it's constructed is a very stable part of the
codebase. As it's likely to change significantly less frequently than
the PanopticonRegisterer itself needs to change, it's fine to rely on
the concrete details inside the class.

The advantage of doing this is that it reduces indirection (it's simpler
to follow what's happening without dependencies being passed in) and it
removes some noise from the wiring file.