Remove use of exceptions for flow control.
Previously, an error was being thrown by the fetchable gem deep within
the services (in the SpecialistDocumentRepository) which was then being
rescued way up the chain in the SpecialistDocumentsController.

By creating a DocumentFinder to be used by the various helpers, which
prevents calling the service and instead returns nil in cases where we
want to render a '404', we're able to remove this use of exceptions.

[+ @bestie]