Remove govspeak lambdas from wiring.
The "SpecialistDocumentGovspeakToHTMLRenderer" and "HeaderExtractor"
classes both already have (in essence) a hard dependency on Govspeak.
They're also not used to render anything other than govspeak.

By pushing the dependency on govspeak down to these classes from the
wiring, and using simple dependency injection for the tests, we can
further simplify the wiring.