Don't check for multiple editions when republishing
We're no longer setting old editions to be archived when we publish a
new one, so raising an error when there's more than one isn't helpful.

Since we've updated the behaviour of `SpecialistDocument#published?` to
behave nicely and we filter the list of documents on that before
republishing them, we don't need to worry about bad data in the form of
multiple published edition any more.