Use PublicationLog for previous published_date
As the repository only pulls out the last 2 document editions there is no
garauntee that we will always have pulled out a document edition that is a
major update. Instead all documents that have been published should have
a PublicationLog entry from this point on. For any that do not have a PublicationLog entry just fall back to the documents updated_at as before.

There is some discussion about pulling more documents from the repository
however we are worried about taking a huge hit on performance which is already an issue here.