Make show view attachments section collapsible
Some documents have a *lot* of attachments. An example of an extreme
case at the moment is the "Local bus services market investigation" from
the CMA, which has 460 attachments.

Since the attachments are all displayed on the show page and the edit,
publish and withdraw buttons are all right at the bottom of the page,
that means editors are having to do way more scrolling than is
reasonable.

To that end:

- Wrap the attachment table in a collapsible `div` and make the heading
  toggle it.
- Add the attachment count to the heading
- Don't make it collapsible where there are no attachments