Duplicate 'org' services in vanilla registries
For Manual services and Manual Document services there are currently 2 types of
Service Registry: a vanilla registry and an organisational registry.

In general, the organisational registry is (unfortunately) more likely to be
dependent on being used from a controller. While the non-organisational
registry might've been created to host services that were being used in scripts
or workers.

Until the day comes that we can pay back the technical debt by possibly
consolidating these two types of registry (and just injecting the repository it
uses??), there are situations where the vanilla registry will need to make use
of those controller-oriented services that are currently exclusive to the
organisational registry.