Add script to rename a document slug
`bin/rename_slug` renames a slug by withdrawing it so that the original
page returns a `410 Gone`. It then republishes the document with the
new slug.

The process is taken from the ops-manual, which currently has
out-of-date code examples.

https://github.gds/pages/gds/opsmanual/2nd-line/changing-specialist-publ
isher-document-slug.html

Example:

    bin/rename_slug "50a02899-0fa9-44d0-aff5-30adeb9898d4"
"cma-cases/new-slug-for-case"

Notes:

- The script is mostly cargo-culted from existing scripts in the repo.
I'm assuming this is how things are supposed to be done in this app.
- This method is not really safe, if republishing or
`update_attributes!` fails, the document will stay `archived`. In that
case you'll have to manually republish.