Remove email alert filtering by category for UTAAC decisions
There is a limitation in the length of the URL sent to the
email configuration service. If too many categories are
selected it is possible to cause an error due to the URL length.

Decision made to remove the category filtering in order to
prevent this potential error from occuring.