Add script and class to handle manual section reslugging
We add a script that calls a class for reslugging manual sections. We first withdraw the old section which places a `gone_item` into the content-store. Then we change the slug of the manual section in the local database. Finally we publish the manual which the section belongs to.

The workflow is inspired by https://github.gds/pages/gds/opsmanual/2nd-line/changing-specialist-publisher-document-slug.html

This PR addresses https://trello.com/c/6zdENr8F/93-re-slug-manual-section

Usage:
`./reslug_manual_section guidance/countryside-stewardship-manual 8-terms-and-conditions 8-scheme-requirements-and-procedures`