Publish a new special document edition for the slug change
Previously, the last edition for a manual section was updated, thereby inheriting its change notes which could be confusing for anyone finding this change.
This commit implements the publishing of a new edition for the manual section with appropriate change_note before republishing.
The workflow is inspired by how the manual_documents_controller handles the update to a section. The current architecture requires initialising an instance of  `ManualDocumentServiceRegistry`, passing it the controller as an argument (which contains a params hash for the updated section). This is mimicked by the dummy object `FakeController` in this commit.