Bump gds-api-adapters for more detailed errors
There's an error when it tries to deploy to the integration environment which
this will help me investigate. Also: better errors are better.

Since 27.1.0, gds-api-adapters includes the request body in the error message,
allowing you to see much more detail by default. This is particularly helpful
when trying to debug pushing to the publishing-api.

I've bumped to the latest gds-api-adapters because there's no reason not to,
and it includes some useful changes. I had to bump plek because
gds-api-adapters needed me to.