Remove the Publishing API-related tasks
As far as I understand, these tasks publish content for the first time
to the Publishing API. They have already been run on production and
any republishing to Publishing API is supposed to be done with the
scripts in `bin` folder.

These tasks should be deleted to avoid confusion with what's in `bin`.
I've also deleted the classes they make direct calls to, because I couldn't
find any other references to them (apart from the specs).