Create a script to delete duplicate drafts
There is a bug that causes duplicate draft editions to be created in
the SP database. These duplicates are rejected from the Publishing API
but cause confusion for publishers, because they suddenly see many
identical drafts that they are unable to delete.

We are not fixing the underlying issue because this app will be replaced
very shortly; we want to remove duplicates to improve the user experience
for publishers and to make republishing to Publishing API possible.

The script works by finding slugs with multiple editions, and then
deletes those editions that aren't present in Publishing API.