Cache specialist documents for maximum 10 seconds
We needed to ensure that specialist documents had a short enough cache time so
that recipients of alert emails would see current content if they looked at
the website.  This change adds the `max_cache_time` onto the details hash so
that it matches the work done in the [specialist-publisher-rebuild](https://github.com/alphagov/specialist-publisher-rebuild/pull/751).

We have set the cache time to 10 seconds as this matches what was originally
done in [Travel Advice Publisher](https://github.com/alphagov/travel-advice-publisher/pull/126).

[Trello card](https://trello.com/c/WGlWkWT5/70-reduce-cache-on-specialist-publisher-medium)