Explicitly document what JSON payload looks like when republishing
Opted for using `request_json_matching` because I want to show that this
is the full body of what is being sent. However, this matcher does not
handle nested hashes well, which is why publishing_api_fields contain
String keys in the nested hash. (An update to gds-api-adapters is
considered not valuable as this is a soon to be deprecated version of
adapters as more apps move to use publishing-api-v2 adapters)