Queue republish_all jobs
Due to the large amount of documents, it is useful to speed up the
republish_all action by sending each document to be republished to a
queue.

Also copied pattern of using a separate low-priority queue for bulk
republishing, so to not affect normal publishing that uses the default
queue.