Add script to delete editions with xx prefixed titles
The application is prone to creating multiple duplicate drafts.  We have
a script, bin/delete_duplicate_drafts, for dealing with these, but
sometimes users will manually mark the duplicates for deletion by adding
xx to the start of the title and asking us to delete them.  When this
happens some of the duplicates end up in the publishing-api and so the
duplicate deleter will not remove them as it only deletes those that the
publishing-api knows nothing about.

This script finds all documents with the 'xx' prefix and deletes them
from the db.  It also asks publishing api to discard the draft if it
knows about it.

It's likely to fail if the editions are not in draft state - but by
default the script only prints out information, to really delete the
editions you must run it with a DO_IT environment variable present.