Republish withdrawn documents…
We need to republish all documents from v1 to v2
for when we move over to the rebuilt app. When we
do so, anything that’s ‘withdrawn’ needs to be
‘unpublished’ in v2 and so we need to use the new
v2 ‘unpublish’ mechanism rather than creating a
‘gone’ piece of content and sending that through.

We recently added support to the Publishing API to
support safely republishing content into a ‘gone’
state. It is now possible to transition content
directly from ‘draft’ to ‘unpublished’, without
going through the ‘published’ state that is deemed
unsafe, i.e. it would result in unpublished
content being temporarily live on GOV.UK.

One thing to be wary of in this work is that
documents are correctly republished when there is
a subsequent draft. Documents can be in more than
one state at once (‘withdrawn’ & ‘draft’) and so
the conditional in RepublishDocumentService needs
to separately check if the document is withdrawn
before sending a draft, if there is one.