Increase the timeout to 30 seconds to the publishing API
A particular content item (/cma-cases/31d49965-5826-42ff-ac87-67f305fbb5f9)
has a large number of attachments and the overall time to send this to
the publishing API and the publishing API's time to process against the
V1 API endpoint takes longer then the default 4 seconds currently set in
the GDS-API adapters. This throws a Timeout exception, even though the
publishing API still processes the request.

The V2 endpoint is a lot more responsive, so should not be a problem
with the rebuild of SP.