Update gds-sso to 10.1.0
This update populates the organisation_content_id field
which the rebuild project depends on.

Rather than update govuk_content_models as well, it
seemed safer to enhance the user model with this field.
govuk_content_models have moved on a lot and some of
the Specialist Publisher models have been removed.