Revert "Merge pull request #745 from alphagov/bump-ruby-version"
This reverts commit 609c896180e00ed4d24052a5a1825525ca286c3f, reversing
changes made to a66664d3a3bfd44907650c58bc464375d589a3dc.

The reason for reverting this version bump is mainly that the rails
console could not be opened in a production environment due to an
error that it required the 'test-unit' gem to be present. This seems
very strange given this dependency isn't usually needed for anything
other than the test environment.
Some research hinted that Ruby 2.3 wasn't altogether compatible with
Rails 3.2 and once this branch was deployed for testing the Rails
console could be started without error.
The mongoid configuration file needed a list of hosts rather than a
uri property to work correctly with Ruby 2.1 due to limitations in the
ruby URI library. This is consistent with Specialist Publisher's
original config.