Add class for rebuilding major manual updates
The existing publication logs for manual edition updates appear to be
out by a few seconds from the exported_at timestamp on the edition and
I'm not sure why. In this rebuilding I have assumed that it is acceptable to
use the exact exported_at timestamp for the publication log.