Remove removed_document_ids entries that point to deleted documents
At some point in the past, we fully deleted some sections from a manual
so that they no longer exist.  Unfortunately we left references to their
ids in the removed_document_ids of the manual they used to belong to. This
means that we can't republish that manual, or do other publishing tasks as
publishing does things with the removed_document_ids.

To solve this we run a migration that finishes the job by removing any ids
in removed_document_ids that no longer point to an instance in
SpecialistDocumentEdition.  We have done some manual work to check that the
ids we remove don't still exist in the publishing-api and they don't so our
cleanup is all we need to do.