Fix ManualsController spec to accurately check a document isn't published
The spy setup was incorrect as it allowed `publish` to be called on the
`ManualsController` class, and then asserted that it wasn't.  This isn't
how rails controllers work (the `publish` action is an instance method
for a start) and so we get false positives from this test.

We change the test to spy on the service object returned from the
controller instance's `services` method.