Scope section withdrawal to gds_editor and editors only
We re-use the PermissionChecker#can_withdraw? method but change it
to gds_editor and editor permissions only.  This method isn't used
anywhere else other than the code we've added so this seems ok for
now, but if we ever want different permissions for withdrawing
manuals vs. sections then we'll need to do something else.