Allow removal of sections from published manuals
This mostly already worked.  What we have to add is:

1. marking the manual as a draft in the remove manual document service
2. adding the publishing api manual draft exporter as a listener to the
   same service.

This means we always write a new draft to the publishing-api for the
manual when we remove a section, and this draft won't have the section
in the details hash as a child.  This is something we can test for in
our cucumber features which is nice.