Present change note fields when removing a section
This means that removing a section will create a new draft edition of the
section being removed as we use the update method to persist the change
note details.

As part of this we change what withdraw means for a SpecialistDocument.
It used to be that withdrawal was something we did to published editions
and left drafts alone.  Instead we now also allow withdraw! to work on
draft editions, and change withdrawn? to look at just the latest edition
instead of the most recent non draft.  It's very likely that this muddles
some of the details around the states (for example published? is true if
any edition was ever published, rather than just the most recent).

Nothing else was using the withdraw! or withdrawn? methods so this feels
safe enough.  The change is made because it means the
publishing_api_publisher callback doesn't need to care if the document is
draft or not in order to archive things, and allows us to avoid duplicate
work in other callbacks.

In the publication_logger callback we also look at the removed_documents
to write out PublicationLog entries for these, if they're not withdrawn yet.
This relies on publication_logger callback running before the
publishing_api_publisher callback, but as this is the case for the normal
documents it feels ok to rely on it.

In order to support rebuilding the PublicationLog entries if anything
goes wrong we also stamp the exported_at timestamp on the edition when
we archive them in the publishing_api_publisher callback.  We do this
in a new withdraw_and_mark_as_exported_to_live_publishing_api! method
which is a horrible name, but avoids doing two separate writes to the
db which is what would haappen if we used withdraw! and
mark_as_exported_to_live_publishing_api! separately.
reconstructing PublicationLog entries.

We also change the publication_logger to