Adds migration to tidy up manuals with sections removed already
We've just improved the functionality for removing sections from a manual
so we want to tidy up sections that were removed before this.  This tidy
up makes sure that publishing-api is kept up-to-date by discarding drafts
and unpublishing them, rummager is asked to delete the section, and that
our database has things in the right state.