Use manual records directly rather than repo when republishing
Previously we used to fetch all manuals for republishing via the
repository, but this caused an exception in the case where a removed
document id was not present for any manual. This commit reworks the
republish script to access the manual records directly and therefore
catch the above case in the right place.

The removed document ids should not be a problem after migration [1] is
run.

1. db/migrate/20161107113454_remove_deleted_removed_document_ids_from_manuals.rb