Set publication log timestamp to edition updated time
While running a task to rebuild publication logs for major updates it
was assumed that all non-draft editions had an exported_at time set. But
some were found to have exported_at as nil. By using the updated_at attribute as
a fallback we use the "best guess" for the time of exporting that edition.