Move ManualRecord::Edition to its own collection
These were originally embedded within the manual_records collection but
this means the entire change history of the manual needs to be under the
document size limit for mongo: 16Mb.  For manuals with lots of changes
this isn't always going to be true and this means users are unable to
save their changes, but don't get useful error messages telling them why.

By moving the editions into their own collection, and changing from an
embeds_many/embedded_in relationship to a has_many/belongs_to relationship
we avoid this problem.  Each individual change now needs to be under 16Mb
when saved to Mongo, but this shouldn't be a problem (or if it is, it's a
larger problem for the system as a whole, not just manual recods and their
editions).

To preserve the interface that editing an edition also stamps the manual
record we add our own implementation of touch to the editions.  Mongoid
gets first class support for touch on associations in mongoid 2.6 (this
was backported from mongoid 3), but we're still on mongoid 2.5.x.