Make ManualRecord#latest_edition aware of in-memory editions
It seems that with an embedded relationship doing .order_by on the
editions for a manual_record would look at in-memory editions, whereas
with a separate collection relationship the .order_by does a DB call
and ignores the already loaded editions relationship and thus any
unsaved instances.  Like those created by
ManualRecord#build_draft_edition.

To get round this we memoize ManualRecord#latest_edition - it still
uses .order_by on the editions if the value hasn't been set, but whenever
we call build_draft_edition we set the latest_edition value to be the
draft that we build.  This'll obviously break if we ever build new
editions not via this API, but it doesn't look like there is anywhere
that we do, so we're safe.

We also have to add an after_save callback to the manual record to save
the latest edition if it is present, and if it has changes that need
saving.  We clear it when we do this, just in case.