Create rebuilt manual publication logs in best-guess section order
The order of creation is important because it ultimately affects the order
of updates on the front end.

Our best guess of the document order (which can change throughout the lifetime
of a manual) is to use the order stored on the most recent manual
edition. Any editions that were previously removed are displayed in created at
order.

The result should be that the frontend will show updates in best guess
section order when grouped by day.