Tighten up PublicationLog#change_notes_for
This method would return all publication logs that had the supplied slug
as a prefix to allow us to get the publication logs for a manual section
at the same time as the publication logs for the manual.  However this
didn't account for manuals whose slugs might be a prefix of another
manual.  For example the manual slug 'guidance/test' is a prefix of the
manual slug 'guidance/test-manual'.  This means when asking for the logs
for 'guidance/test' you would get the logs for both manuals (and all their
sections) which is almost certainly not what you wanted.

We tighten up the regexp used to extract the logs so that the slug is
followed by a / (so we get logs for child sections) or the end of the
string (so we get logs for the manual).

This has happened as there are a handful of manuals whose slugs are a
prefix of another manual, but in practice they're all either withdrawn
or unpublished, so there are no logs.  But it's likely to become a bigger
problem if we don't fix it.