Log manual edition document editions at manual timestamp
It was noticed that some publication logs for document editions have an
exported_at timestamp that post-dates the timestamp on the associated
manual edition's updated_at timestamp. This seemed to particularly
affect logs for documents that were added in the first manual edition.

It's possible that various rebuilding/republishing scripts have
modifed document editions and therefore that their timestamps are
less reliable for the log rebuild.

Here we assume that the updated_at timestamp on the first manual edition is
the correct timestamp to use for the publication log for its document
editions.

This commit adds logic to the publication log rebuilder so that it
checks for the first manual edition documents and sets the
associated publication log to the manual edition updated_at time.
Subsequent document editions still have publication logs using the
(possibly inaccurate) document edition exported_at or updated_at timestamp.