Log first document editions at the manual edition timestamp
It was noticed that some publication logs for document editions have an
exported_at timestamp that post-dates the timestamp on the associated
manual edition's updated_at timestamp. The normal workflow should result
in manual editions that are updated at the point when they are
exported/published and are then no longer updated.

It's possible that various rebuilding/republishing scripts have
modifed document editions and therefore that their timestamps are
less reliable for the log rebuild.

Here we assume that the updated_at timestamp on the manual edition is
the correct timestamp to use for the publication log for document
editions. However, we can only infer the association
between manual editions and document editions in the case where they are both at the
first version number. This is because it's not currently possible to ask
a manual edition which versions of documents it relates to.

This commit adds logic to the publication log rebuilder so that it
checks for all first document editions and sets the timestamps on the
associated publication log to the document's manual edition updated_at time.
Subsequent document editions still have publication logs using the
(possibly inaccurate) document edition exported_at or updated_at timestamp.