Refactor ManualRelocator slightly
We're about to add some more functionality, and having an object rather
than a bunch of class methods makes some things easier.

The one change we introduce is that the relocator now won't work on
multiple old documents that have the same slug and will raise an error.

We also add some more specs to make sure we're doing what we think we are
when destroying and reslugging things.