Issue gone unpublishings for slugs we'll reuse
To reuse a path in publishing api you have to turn it into one of the
types that can be substituted.  In our case a gone is the most sensible
option.  We send a gone for the existing manual base path itself as this
path will be re-used by the temporary manual when we re-slug it.  We also
work out if any sections on the existing manual would have the same path
as any sections on the temporary manual once they are reslugged and issue
gones for those existing manual sections.