Put new drafts of the temporary manual and its sections
These are published so that they tell the publishing-api that
these content_ids now have new routes.  We don't use the existing
wiring, repositories and so on because this would tie us into
always sending drafts of the most current editions.  In coming
commits we're going to send drafts+publishes of previous editions
if the temp manual has been published.  The duplication of code
here seems worth while to let us achieve this.

Note that we're not sending new links as I don't think any of
these will have changed because of the reslugging.