Only relocate manuals if both are published
Specifically that the latest_edition is published or the latest_edition
is draft and the previous edition is published.  Other states might work
but we're not going to deal with it.  If the temp manual isn't published
we'll end up deleting content: we get rid of the existing manual with a
gone (so the path can be reused) and the very next thing to do is publish
content from the temp manual back onto that path which it can't do if the
temp manual isn't published.  If the existing manual isn't published we
can't issue a gone so things might break.