Redraft and republish the temporary manual
If the temp manual is published we send the current edition as a draft, and
send the latest edition of all sections as drafts, then send publish events
for all of these.  If it's got a draft we send a draft for the previous
edition of the manual, and the most recent published version of all the
sections.  We then send a publish event for all of these, and finally
write out drafts of the current edition of the manual and all the latest
editions of the sections.

This should get us back into the right shape.