Issue "gones" via draft/publish not unpublish
To leave a base_path in a state where a different content_id can use it
we have to remove the existing manual and its sections via a draft+publish
of type gone, rather than sending an unpublish of type gone.

From talking to the publishing platform team it sounds like this is a bit
of a hack / leftover from before the unpublish end point was added, we
might be able to go back to the unpublish version at some point.