Remove the bad draft in our migration
As well as removing the old capital-funding-guide manual and
reslugging the new one, we also want to remove an bad draft
that has no title.  We use the existing CliManualDeleter for
this purpose, and pass in a StringIO with a "y" in it to get
around the required keyboard input.