Set update_type for manuals based on the documents that need exporting
When we draft a manual the update_type for the publishing-api is based
on the minor_update? state of all the documents that belong to the
manual because the manual editions themselves don't have change notes
and minor/major update states.

There is a bug here because it looks at the minor_update? state of all of
the documents meaning that a major update published several editions ago
can cause each publish event to go through as a major update.  For
example: create a manual with 3 documents and publish it; the state for
all these documents will be that of a major update.  If we edit one of
them with a minor update and publish the manual, the update_type will
still be major when we've really only published a minor update.  We'll
only get a minor update when we've edited and published all 3 of the
documents with minor update change notes.  If we only look at those
documents that need exporting we get the correct behaviour, so that's
what we do.