Update ManualRelocator to understand about sections that need_exporting
This means the ManualRelocator can work again when publishing as the
update_type attribute now relies on the needs_exporting? attribute of
the sections.  We can't detect the update_type correctly for the previous
draft (it'll always be minor as none of the section editions will need
exporting) but that can't be helped.