Stop sending publishing api public_updated_at
Prior to this change the app would send public_updated_at to the
publishing-api filled with the updated_at value of the manual or
section being drafted.  This isn't the right value to use because
we write drafts at edit time, not publish time.  So when we publish
the timestamp shown to the user will be the edit time.

It turns out that Publishing API will set timestamps correctly based
on the api calls made if the dates aren't already set, and if the
update_types on the put_content and publish events are set correctly (see
https://github.com/alphagov/publishing-api/blob/master/doc/api.md).
We've just changed it so that the update_types are set correctly which
means we can stop sending public_updated_at and get the correct timestamps
set (at publication time, not at draft time).