Add has_ever_been_published? methods to the document models
This lets us ask if the document has ever been published before, not just
if the current edition is published.  For ManualRecord this is true if
any of the editions have the state published and we pass this value on to
the Manual instance via the repository when building a Manual.  For
SpecialistDocument (which acts as the Manual and ManualRecord equivalents
for SpecialistDocumentEditions) we alias the existing published? method.

Having this consistent interface across the document types will come
in useful in the next commit where we have to change the update_type
if the document has never been published before.