Set update_type to "major" for first editions sent to publishing-api
The publishing-api will complain if the first version of a content item
does not have update_type of "major".  At the moment this causes a 500
error when you publish but it will eventually turn into a proper error
and rejection (either at drafting or publishing).  This means we need
to make sure that the first editions of a manual or section are sent as
"major" updates.

Calculating if this is the first time something has been sent to the
publishing-api isn't very simple however as we don't currently track this
on manuals, only sections (via the exported_at attribute).  We can't use
the version_number as this doesn't increment each time something is
published, only each time it is changed.  For example it is possible that
v1 of the manual is the current version for all publish events as all the
changes happen in the sections.  We could combine the version_number
and the state (for example version_number: 1, and state: draft) but the
state is changed before we export to the publishing API so we'd never
see a manual that does not have a state of 'published'.  We fall back
to checking if the manual has ever been published (does it have any
published editions at the time it is fetched from the repo) and deciding
based on that if the update is major or if we should ask the documents.

We do the same thing with the documents - we check if they have ever been
published as well as checking if any are major.