Avoid flappy test in manual_publication_log_filter_spec
This test would fail because it expects the PublicationLogs to be
returned in a specific order, but our test data can't guarantee that
order because it's based on timestamp ordering and everything has
the same timestamps.  We can assert however what the first 3 should
be and what the last 2 should be so we do that using two match_array
calls which doesn't care about order.