Allow editing of change notes on all sections
We want to allow change notes on manual sections to be added and edited
throughout the process of editing a manual.  Prior to this change a change
note can only be added to the first section you edit for a manual, and
only on that initial edit.  If you want to a a change not to another
section you have to publish the outstanding changes and draft the manual
afresh.  You weren't then able to change the note you added to the
first section either.

This is obviously not a great workflow, so this commit changes it so
that the only restriction is that you can't add a change note to the
sections you add to the very first draft edition of a manual (this seems
to be the intention), but after that first draft is published you can
add change notes to every section you add or edit and you can go back
to edit them inbetween saves.  When you publish all notes are cleared
again and you can't edit the old ones (but you can add new ones).